Optimize the framework: (1) now taskId is an integer; (2) add treeId to label the sub-protocols; (3) adjust AbstractMultiPartyPto.java so that developers do not need to manually override functions like setParallel, setTaskId.

+++ mpc4j-s2pc-pcg/src/test/java/edu/alibaba/mpc4j/s2pc/pcg/mtg/zl/core/ZlCoreMtgTest.java
@@ -152,7 +152,7 @@ public class ZlCoreMtgTest {
     }
 
     private void testPto(ZlCoreMtgParty sender, ZlCoreMtgParty receiver, int num) {
-        long randomTaskId = Math.abs(SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(SECURE_RANDOM.nextInt());
         sender.setTaskId(randomTaskId);
         receiver.setTaskId(randomTaskId);
         try {
