+++ mpc4j-s2pc-pso/src/test/java/edu/alibaba/mpc4j/s2pc/pso/oprf/MpOprfTest.java
@@ -157,7 +157,7 @@ public class MpOprfTest {
     }
 
     private void testPto(MpOprfSender sender, MpOprfReceiver receiver, int batchSize) {
-        long randomTaskId = Math.abs(SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(SECURE_RANDOM.nextInt());
         sender.setTaskId(randomTaskId);
         receiver.setTaskId(randomTaskId);
         try {
