+++ mpc4j-sml-opboost/src/test/java/edu/alibaba/mpc4j/sml/opboost/xgboost/OpXgBoostRegSingleSlaveTest.java
@@ -207,7 +207,7 @@ public class OpXgBoostRegSingleSlaveTest {
     private Predictor federateTraining(Formula formula,
                                        DataFrame hostDataFrame, OpXgBoostHostConfig hostConfig,
                                        DataFrame slaveDataFrame, OpBoostSlaveConfig slaveConfig) {
-        long randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextLong());
+        int randomTaskId = Math.abs(OpBoostTestUtils.SECURE_RANDOM.nextInt());
         host.setTaskId(randomTaskId);
         slave.setTaskId(randomTaskId);
         try {
